feat: service account log
* The config loader will first try to find a ~/.kube/config. If that fails, then check for the service account files

fixes #1
